finish fastapi authorization

+++ routers/posts.py
@@ -4,7 +4,7 @@ from .. import models
 from sqlalchemy.ext.asyncio import AsyncSession
 from sqlalchemy.orm import selectinload
 from sqlmodel import col, select
-
+from ..auth import CurrentUser
 
 router = APIRouter()
 
@@ -14,19 +14,12 @@ router = APIRouter()
     response_model=models.PostResponse,
     status_code=status.HTTP_201_CREATED
 )
-async def create_post(post: models.PostCreate, db: Annotated[AsyncSession, Depends(models.get_db)]):
-    result = await db.execute(select(models.User).where(col(models.User.id) == post.user_id))
-    user = result.scalars().first()
-    if not user:
-        raise HTTPException(
-            status_code=status.HTTP_404_NOT_FOUND,
-            detail="User not found",
-        )
+async def create_post(post: models.PostCreate, current_user: CurrentUser, db: Annotated[AsyncSession, Depends(models.get_db)]):
 
     new_post = models.Post(
         title=post.title,
         content=post.content,
-        user_id=post.user_id,
+        user_id=current_user.id
     )
     db.add(new_post)
     await db.commit()
@@ -52,22 +45,17 @@ async def get_post(post_id: int, db: Annotated[AsyncSession, Depends(models.get_
 
 
 @router.put("/{post_id}", response_model=models.PostResponse)
-async def update_post_full(post_id: int, post_data: models.PostCreate, db: Annotated[AsyncSession, Depends(models.get_db)]):
+async def update_post_full(post_id: int, post_data: models.PostCreate, current_uer: CurrentUser, db: Annotated[AsyncSession, Depends(models.get_db)]):
     result = await db.execute(select(models.Post).where(col(models.Post.id) == post_id))
     post = result.scalars().first()
     if not post:
         raise HTTPException(status_code=status.HTTP_404_NOT_FOUND, detail="Post not found")
-    if post_data.user_id != post.user_id:
-        result = await db.execute(select(models.User).where(col(models.User.id) == post_data.user_id))
-        user = result.scalars().first()
-        if not user:
-            raise HTTPException(
-            status_code=status.HTTP_404_NOT_FOUND,
-            detail="User not found",
-        )
+    
+    if post.user_id != current_uer.id:
+        raise HTTPException(status_code=status.HTTP_403_FORBIDDEN, detail="Not authorized to update this post")
+
     post.title = post_data.title
     post.content = post_data.content
-    post.user_id = post_data.user_id
 
     await db.commit()
     await db.refresh(post, attribute_names=["author"])
@@ -75,11 +63,14 @@ async def update_post_full(post_id: int, post_data: models.PostCreate, db: Annot
 
 
 @router.patch("/{post_id}", response_model=models.PostResponse)
-async def update_post_partial(post_id: int, post_data: models.PostUpdate, db: Annotated[AsyncSession, Depends(models.get_db)]):
+async def update_post_partial(post_id: int, post_data: models.PostUpdate, current_user: CurrentUser, db: Annotated[AsyncSession, Depends(models.get_db)]):
     result = await db.execute(select(models.Post).where(col(models.Post.id) == post_id))
     post = result.scalars().first()
     if not post:
         raise HTTPException(status_code=status.HTTP_404_NOT_FOUND, detail="Post not found")
+    
+    if post.user_id != current_user.id:
+        raise HTTPException(status_code=status.HTTP_403_FORBIDDEN, detail="Not authorized to update this post")
 
     update_data = post_data.model_dump(exclude_unset=True)
     for field, value in update_data.items():
@@ -92,11 +83,14 @@ async def update_post_partial(post_id: int, post_data: models.PostUpdate, db: An
 
 
 @router.delete("/{post_id}", status_code=status.HTTP_204_NO_CONTENT)
-async def delete_post(post_id: int, db: Annotated[AsyncSession, Depends(models.get_db)]):
+async def delete_post(post_id: int, current_user: CurrentUser, db: Annotated[AsyncSession, Depends(models.get_db)]):
     result = await db.execute(select(models.Post).where(col(models.Post.id) == post_id))
     post = result.scalars().first()
     if not post:
         raise HTTPException(status_code=status.HTTP_404_NOT_FOUND, detail="Post not found")
+    
+    if post.user_id != current_user.id:
+        raise HTTPException(status_code=status.HTTP_403_FORBIDDEN, detail="Not authorized to delete this post")
 
     await db.delete(post)
     await db.commit()
